for x86, but consider that musl lacks setcontext implementation. so it cannot be compiled on x86
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@36687 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/rules/libunwind.mk
@@ -15,16 +15,21 @@ libunwind-configure:
 	sed -i 's/need_relink=yes/need_relink=no/g' $(TOP)/libunwind/libtool
 
 libunwind: 
+ifeq ($(ARCH),i386)
 	$(MAKE) -C libunwind
+endif
 
 libunwind-clean: 
+ifeq ($(ARCH),i386)
 	if test -e "libunwind/Makefile"; then $(MAKE) -C libunwind clean ; fi
+endif
 
 libunwind-install: 
+ifeq ($(ARCH),i386)
 	$(MAKE) -C libunwind install DESTDIR=$(INSTALLDIR)/libunwind
 	rm -rf $(INSTALLDIR)/libunwind/usr/man
 	rm -rf $(INSTALLDIR)/libunwind/usr/include
 	rm -f $(INSTALLDIR)/libunwind/usr/lib/*.a
 	rm -f $(INSTALLDIR)/libunwind/usr/lib/*.la
 	rm -rf $(INSTALLDIR)/libunwind/usr/lib/pkgconfig
-	
+endif
